backend/requirements.txt
fastapi
pydantic-settings
alembic
tenacity
psycopg2-binary
passlib
python-jose
emails
